moved generate_unique_codename to source_app/utils.py
