+++ securedrop/source.py
@@ -22,7 +22,8 @@ import store
 from db import db_session, Source, Submission, Reply, get_one_or_else
 from source_app import create_app
 from source_app.decorators import login_required, ignore_static
-from source_app.utils import logged_in, valid_codename
+from source_app.utils import (logged_in, valid_codename,
+                              generate_unique_codename)
 
 import logging
 # This module's logger is explicitly labeled so the correct logger is used,
@@ -84,31 +85,6 @@ def check_tor2web():
             'This <strong>does not</strong> provide anonymity. '
             '<a href="/tor2web-warning">Why is this dangerous?</a>')),
               "banner-warning")
-
-
-def generate_unique_codename():
-    """Generate random codenames until we get an unused one"""
-    while True:
-        codename = crypto_util.genrandomid(Source.NUM_WORDS)
-
-        # The maximum length of a word in the wordlist is 9 letters and the
-        # codename length is 7 words, so it is currently impossible to
-        # generate a codename that is longer than the maximum codename length
-        # (currently 128 characters). This code is meant to be defense in depth
-        # to guard against potential future changes, such as modifications to
-        # the word list or the maximum codename length.
-        if len(codename) > Source.MAX_CODENAME_LEN:
-            app.logger.warning(
-                    "Generated a source codename that was too long, "
-                    "skipping it. This should not happen. "
-                    "(Codename='{}')".format(codename))
-            continue
-
-        filesystem_id = crypto_util.hash_codename(codename)  # scrypt (slow)
-        matching_sources = Source.query.filter(
-            Source.filesystem_id == filesystem_id).all()
-        if len(matching_sources) == 0:
-            return codename
 
 
 @app.route('/generate', methods=('GET', 'POST'))
